TB_CPA_Harmonize_dev_beta/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
pyyaml>=6.0
tqdm>=4.65.0
tabulate>=0.9.0
python-magic-bin>=0.4.14; sys_platform == 'win32'
python-magic>=0.4.27; sys_platform != 'win32'
chardet>=5.0.0
